% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_bundle_instance}
\alias{ec2_bundle_instance}
\title{Bundles an Amazon instance store-backed Windows instance}
\usage{
ec2_bundle_instance(InstanceId, Storage, DryRun)
}
\arguments{
\item{InstanceId}{[required] The ID of the instance to bundle.

Type: String

Default: None

Required: Yes}

\item{Storage}{[required] The bucket in which to store the AMI. You can specify a bucket that you
already own or a new bucket that Amazon EC2 creates on your behalf. If
you specify a bucket that belongs to someone else, Amazon EC2 returns an
error.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}
}
\description{
Bundles an Amazon instance store-backed Windows instance.
}
\details{
During bundling, only the root device volume (C:\) is bundled. Data on
other instance store volumes is not preserved.

This action is not applicable for Linux/Unix instances or Windows
instances that are backed by Amazon EBS.
}
\section{Request syntax}{
\preformatted{svc$bundle_instance(
  InstanceId = "string",
  Storage = list(
    S3 = list(
      AWSAccessKeyId = "string",
      Bucket = "string",
      Prefix = "string",
      UploadPolicy = raw,
      UploadPolicySignature = "string"
    )
  ),
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
