% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_cancel_import_task}
\alias{ec2_cancel_import_task}
\title{Cancels an in-process import virtual machine or import snapshot task}
\usage{
ec2_cancel_import_task(CancelReason, DryRun, ImportTaskId)
}
\arguments{
\item{CancelReason}{The reason for canceling the task.}

\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{ImportTaskId}{The ID of the import image or import snapshot task to be canceled.}
}
\description{
Cancels an in-process import virtual machine or import snapshot task.
}
\section{Request syntax}{
\preformatted{svc$cancel_import_task(
  CancelReason = "string",
  DryRun = TRUE|FALSE,
  ImportTaskId = "string"
)
}
}

\keyword{internal}
