% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_export_tasks}
\alias{ec2_describe_export_tasks}
\title{Describes one or more of your export tasks}
\usage{
ec2_describe_export_tasks(ExportTaskIds)
}
\arguments{
\item{ExportTaskIds}{One or more export task IDs.}
}
\description{
Describes one or more of your export tasks.
}
\section{Request syntax}{
\preformatted{ec2$describe_export_tasks(
  ExportTaskIds = list(
    "string"
  )
)
}
}

\keyword{internal}
