% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_describe_import_snapshot_tasks}
\alias{ec2_describe_import_snapshot_tasks}
\title{Describes your import snapshot tasks}
\usage{
ec2_describe_import_snapshot_tasks(DryRun, Filters, ImportTaskIds,
  MaxResults, NextToken)
}
\arguments{
\item{DryRun}{Checks whether you have the required permissions for the action, without
actually making the request, and provides an error response. If you have
the required permissions, the error response is \code{DryRunOperation}.
Otherwise, it is \code{UnauthorizedOperation}.}

\item{Filters}{One or more filters.}

\item{ImportTaskIds}{A list of import snapshot task IDs.}

\item{MaxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{NextToken}
value.}

\item{NextToken}{A token that indicates the next page of results.}
}
\description{
Describes your import snapshot tasks.
}
\section{Request syntax}{
\preformatted{svc$describe_import_snapshot_tasks(
  DryRun = TRUE|FALSE,
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  ImportTaskIds = list(
    "string"
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
