% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_get_host_reservation_purchase_preview}
\alias{ec2_get_host_reservation_purchase_preview}
\title{Preview a reservation purchase with configurations that match those of
your Dedicated Host}
\usage{
ec2_get_host_reservation_purchase_preview(HostIdSet, OfferingId)
}
\arguments{
\item{HostIdSet}{[required] The IDs of the Dedicated Hosts with which the reservation is associated.}

\item{OfferingId}{[required] The offering ID of the reservation.}
}
\description{
Preview a reservation purchase with configurations that match those of
your Dedicated Host. You must have active Dedicated Hosts in your
account before you purchase a reservation.
}
\details{
This is a preview of the PurchaseHostReservation action and does not
result in the offering being purchased.
}
\section{Request syntax}{
\preformatted{ec2$get_host_reservation_purchase_preview(
  HostIdSet = list(
    "string"
  ),
  OfferingId = "string"
)
}
}

\keyword{internal}
