% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2_operations.R
\name{ec2_unassign_ipv_6_addresses}
\alias{ec2_unassign_ipv_6_addresses}
\title{Unassigns one or more IPv6 addresses from a network interface}
\usage{
ec2_unassign_ipv_6_addresses(Ipv6Addresses, NetworkInterfaceId)
}
\arguments{
\item{Ipv6Addresses}{[required] The IPv6 addresses to unassign from the network interface.}

\item{NetworkInterfaceId}{[required] The ID of the network interface.}
}
\description{
Unassigns one or more IPv6 addresses from a network interface.
}
\section{Request syntax}{
\preformatted{ec2$unassign_ipv_6_addresses(
  Ipv6Addresses = list(
    "string"
  ),
  NetworkInterfaceId = "string"
)
}
}

\keyword{internal}
