% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_discover_poll_endpoint}
\alias{ecs_discover_poll_endpoint}
\title{This action is only used by the Amazon ECS agent, and it is not intended
for use outside of the agent}
\usage{
ecs_discover_poll_endpoint(containerInstance, cluster)
}
\arguments{
\item{containerInstance}{The container instance ID or full ARN of the container instance. The ARN
contains the \code{arn:aws:ecs} namespace, followed by the Region of the
container instance, the AWS account ID of the container instance owner,
the \code{container-instance} namespace, and then the container instance ID.
For example,
\code{arn:aws:ecs:<i>region</i>:<i>aws_account_id</i>:container-instance/<i>container_instance_ID</i> }.}

\item{cluster}{The short name or full Amazon Resource Name (ARN) of the cluster to
which the container instance belongs.}
}
\description{
This action is only used by the Amazon ECS agent, and it is not intended
for use outside of the agent.
}
\details{
Returns an endpoint for the Amazon ECS agent to poll for updates.
}
\section{Request syntax}{
\preformatted{svc$discover_poll_endpoint(
  containerInstance = "string",
  cluster = "string"
)
}
}

\keyword{internal}
