% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_operations.R
\name{ecs_update_task_set}
\alias{ecs_update_task_set}
\title{Modifies a task set}
\usage{
ecs_update_task_set(cluster, service, taskSet, scale)
}
\arguments{
\item{cluster}{[required] The short name or full Amazon Resource Name (ARN) of the cluster that
hosts the service that the task set exists in.}

\item{service}{[required] The short name or full Amazon Resource Name (ARN) of the service that
the task set exists in.}

\item{taskSet}{[required] The short name or full Amazon Resource Name (ARN) of the task set to
update.}

\item{scale}{[required]}
}
\description{
Modifies a task set. This is used when a service uses the \code{EXTERNAL}
deployment controller type. For more information, see \href{http://docs.aws.amazon.com/AmazonECS/latest/developerguide/deployment-types.html}{Amazon ECS Deployment Types}
in the \emph{Amazon Elastic Container Service Developer Guide}.
}
\section{Request syntax}{
\preformatted{ecs$update_task_set(
  cluster = "string",
  service = "string",
  taskSet = "string",
  scale = list(
    value = 123.0,
    unit = "PERCENT"
  )
)
}
}

\keyword{internal}
