% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_disk}
\alias{lightsail_delete_disk}
\title{Deletes the specified block storage disk}
\usage{
lightsail_delete_disk(diskName)
}
\arguments{
\item{diskName}{[required] The unique name of the disk you want to delete (e.g., \code{my-disk}).}
}
\description{
Deletes the specified block storage disk. The disk must be in the
\code{available} state (not attached to a Lightsail instance).
}
\details{
The disk may remain in the \code{deleting} state for several minutes.

The \code{delete disk} operation supports tag-based access control via
resource tags applied to the resource identified by diskName. For more
information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\section{Request syntax}{
\preformatted{lightsail$delete_disk(
  diskName = "string"
)
}
}

\keyword{internal}
