% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_instance}
\alias{lightsail_delete_instance}
\title{Deletes a specific Amazon Lightsail virtual private server, or
\emph{instance}}
\usage{
lightsail_delete_instance(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the instance to delete.}
}
\description{
Deletes a specific Amazon Lightsail virtual private server, or
\emph{instance}.
}
\details{
The \code{delete instance} operation supports tag-based access control via
resource tags applied to the resource identified by instanceName. For
more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\section{Request syntax}{
\preformatted{lightsail$delete_instance(
  instanceName = "string"
)
}
}

\keyword{internal}
