% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_delete_instance_snapshot}
\alias{lightsail_delete_instance_snapshot}
\title{Deletes a specific snapshot of a virtual private server (or \emph{instance})}
\usage{
lightsail_delete_instance_snapshot(instanceSnapshotName)
}
\arguments{
\item{instanceSnapshotName}{[required] The name of the snapshot to delete.}
}
\description{
Deletes a specific snapshot of a virtual private server (or \emph{instance}).
}
\details{
The \code{delete instance snapshot} operation supports tag-based access
control via resource tags applied to the resource identified by
instanceSnapshotName. For more information, see the \href{https://lightsail.aws.amazon.com/ls/docs/en/articles/amazon-lightsail-controlling-access-using-tags}{Lightsail Dev Guide}.
}
\section{Request syntax}{
\preformatted{svc$delete_instance_snapshot(
  instanceSnapshotName = "string"
)
}
}

\keyword{internal}
