% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_active_names}
\alias{lightsail_get_active_names}
\title{Returns the names of all active (not deleted) resources}
\usage{
lightsail_get_active_names(pageToken)
}
\arguments{
\item{pageToken}{A token used for paginating results from your get active names request.}
}
\description{
Returns the names of all active (not deleted) resources.
}
\section{Request syntax}{
\preformatted{svc$get_active_names(
  pageToken = "string"
)
}
}

\keyword{internal}
