% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_bundles}
\alias{lightsail_get_bundles}
\title{Returns the list of bundles that are available for purchase}
\usage{
lightsail_get_bundles(includeInactive, pageToken)
}
\arguments{
\item{includeInactive}{A Boolean value that indicates whether to include inactive bundle
results in your request.}

\item{pageToken}{A token used for advancing to the next page of results from your get
bundles request.}
}
\description{
Returns the list of bundles that are available for purchase. A bundle
describes the specs for your virtual private server (or \emph{instance}).
}
\section{Request syntax}{
\preformatted{svc$get_bundles(
  includeInactive = TRUE|FALSE,
  pageToken = "string"
)
}
}

\keyword{internal}
