% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_disk}
\alias{lightsail_get_disk}
\title{Returns information about a specific block storage disk}
\usage{
lightsail_get_disk(diskName)
}
\arguments{
\item{diskName}{[required] The name of the disk (e.g., \code{my-disk}).}
}
\description{
Returns information about a specific block storage disk.
}
\section{Request syntax}{
\preformatted{svc$get_disk(
  diskName = "string"
)
}
}

\keyword{internal}
