% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_domains}
\alias{lightsail_get_domains}
\title{Returns a list of all domains in the user's account}
\usage{
lightsail_get_domains(pageToken)
}
\arguments{
\item{pageToken}{A token used for advancing to the next page of results from your get
domains request.}
}
\description{
Returns a list of all domains in the user's account.
}
\section{Request syntax}{
\preformatted{svc$get_domains(
  pageToken = "string"
)
}
}

\keyword{internal}
