% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_instance_port_states}
\alias{lightsail_get_instance_port_states}
\title{Returns the port states for a specific virtual private server, or
\emph{instance}}
\usage{
lightsail_get_instance_port_states(instanceName)
}
\arguments{
\item{instanceName}{[required] The name of the instance.}
}
\description{
Returns the port states for a specific virtual private server, or
\emph{instance}.
}
\section{Request syntax}{
\preformatted{svc$get_instance_port_states(
  instanceName = "string"
)
}
}

\keyword{internal}
