% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_key_pairs}
\alias{lightsail_get_key_pairs}
\title{Returns information about all key pairs in the user's account}
\usage{
lightsail_get_key_pairs(pageToken)
}
\arguments{
\item{pageToken}{A token used for advancing to the next page of results from your get key
pairs request.}
}
\description{
Returns information about all key pairs in the user's account.
}
\section{Request syntax}{
\preformatted{svc$get_key_pairs(
  pageToken = "string"
)
}
}

\keyword{internal}
