% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_static_ip}
\alias{lightsail_get_static_ip}
\title{Returns information about a specific static IP}
\usage{
lightsail_get_static_ip(staticIpName)
}
\arguments{
\item{staticIpName}{[required] The name of the static IP in Lightsail.}
}
\description{
Returns information about a specific static IP.
}
\section{Request syntax}{
\preformatted{svc$get_static_ip(
  staticIpName = "string"
)
}
}

\keyword{internal}
