% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightsail_operations.R
\name{lightsail_get_static_ips}
\alias{lightsail_get_static_ips}
\title{Returns information about all static IPs in the user's account}
\usage{
lightsail_get_static_ips(pageToken)
}
\arguments{
\item{pageToken}{A token used for advancing to the next page of results from your get
static IPs request.}
}
\description{
Returns information about all static IPs in the user's account.
}
\section{Request syntax}{
\preformatted{lightsail$get_static_ips(
  pageToken = "string"
)
}
}

\keyword{internal}
