% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverlessapplicationrepository_operations.R
\name{serverlessapplicationrepository_get_application_policy}
\alias{serverlessapplicationrepository_get_application_policy}
\title{Retrieves the policy for the application}
\usage{
serverlessapplicationrepository_get_application_policy(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}
}
\description{
Retrieves the policy for the application.
}
\section{Request syntax}{
\preformatted{serverlessapplicationrepository$get_application_policy(
  ApplicationId = "string"
)
}
}

\keyword{internal}
